feat: Support configurable srcs (srcs with select)
Sometimes you want to act on different input files depending on the
platform or some string flag you've defined. Prior to this change,
that would fail, because you can't call len() on a "select", or for
that matter inspect it using the index operator. With this patch, we
skip the check for .json or .xml files if we encounter a "select" in
the srcs attribute; if the user wants both .json/.xml and configurable
attributes, they need to add the appropriate flag to args themselves.