feat: Support Label objects in srcs (#18)
The srcs argument to yq is supposed to be a list of input file labels,
but it did not actually support Label objects, only labels in their
string form. This patch adds support for Lable objects in srcs.

This is useful in particular when yq is called from a macro, as "outer
macros should always prefer to pass Label objects to inner macros
rather than label strings" (according to the documentation of
package_relative_label). Symbolic macros also steer the macro writer
towards using Label objects over label strings.